fix(tests): fold the equivalence proof into its mirrored test file
`test__a2a_host_equivalence.py` had no `_a2a_host_equivalence.py` beside it in
src/, which is an orphan test file under PS-204 §2 — tests mirror the source
layout exactly. CI caught it on all three Python versions; develop's last five
runs are green, so the violation was introduced here.

The tests move into `test__a2a_host_line.py`, which is where they belonged:
they assert the editor's zero-behaviour-change claim, and that claim is
`_a2a_host_line`'s. No test was dropped or weakened — 39 pass across the two
config files (27 + 12), and `audit-project` now reports 1 error where it
reported 2, the survivor being a pre-existing PS-224 on `.github/workflows/
cla.yml`, a file this branch never touches.
